SDL2 interface, on Mac OS X

+++ src/pager.c
@@ -581,7 +581,7 @@ const char **firstmatch;
      * "the interior of a monster".
      */
     if (u.uswallow && looked
-        && (is_swallow_sym(sym) || (int) showsyms[S_stone] == sym)) {
+        && (is_swallow_sym(sym) || showsyms[S_stone] == sym)) {
         if (!found) {
             Sprintf(out_str, "%s%s", prefix, mon_interior);
             *firstmatch = mon_interior;
